requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
pydantic>=2.6,<3
pyyaml>=6.0.1,<7
matplotlib>=3.8,<4
pytest>=8,<9
